refactor(create): replace site nav with a builder top bar
