+++ www/src/modules/create/builder/index.ts
@@ -1,1 +1,2 @@
 export { BuilderPanel } from './builder-panel'
+export { CreateTopBar } from './top-bar'
